Refactoring the code creating datastore package
